Implements Navigation Bar, Refactors Page Navigation
