+++ lib/main.dart
@@ -11,11 +11,11 @@ import 'package:firebase_database/firebase_database.dart';
 import 'package:bullet/firebase_options.dart';
 
 import 'objects/PostCard.dart';
-import 'dart:math';
 
 part 'gui/login.dart';
 part 'gui/signup.dart';
 part 'gui/home.dart';
+part 'gui/profile.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -52,20 +52,60 @@ class MyAppPage extends StatefulWidget {
 
 class _MyAppPageState extends State<MyAppPage> {
   bool isLoggedIn = false;
+  int selectedIndex = 0;
 
   @override
   void initState() {
     super.initState();
     FirebaseAuth.instance.authStateChanges().listen((User? user) {
       isLoggedIn = (user != null);
-      if (mounted) {
-        setState(() {});
-      }
+      setState(() {
+      });
     });
   }
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(body: isLoggedIn ? HomePage() : LoginPage());
+    Widget page;
+    switch (selectedIndex) {
+      case 0:
+        page = HomePage();
+        break;
+      case 1:
+        page = ProfilePage();
+        break;
+      default:
+        throw UnimplementedError('no widget for $selectedIndex');
+    }
+
+    return Scaffold(
+      body: isLoggedIn ? page : LoginPage(),
+      bottomNavigationBar: isLoggedIn ? AppNavigationBar() : SizedBox.shrink(),
+    );
+  }
+
+  AppNavigationBar(){
+    return BottomNavigationBar(
+      backgroundColor: Color.fromRGBO(27, 26, 25, 1),
+      iconSize: 28,
+      items: const <BottomNavigationBarItem>[
+        BottomNavigationBarItem(
+          icon: Icon(Icons.view_list_outlined),
+          label: 'Threads',
+        ),
+        BottomNavigationBarItem(
+          icon: Icon(Icons.person),
+          label: 'Profile',
+        ),
+      ],
+      currentIndex: selectedIndex,
+      unselectedItemColor: Color.fromRGBO(100, 100, 100, 1),
+      selectedItemColor: Colors.white,
+      onTap: (index){
+        setState(() {
+          selectedIndex = index;
+        });
+      },
+    );
   }
 }
